Make terraform an optional dependency
