+++ pyproject.toml
@@ -65,7 +65,6 @@ pymysql = { version = "~1.0.2"}
 pyparsing = "<3,>=2.4.0"
 python = ">=3.7.1,<3.11"
 python-dateutil = "^2.8.1"
-python-terraform = { version = "^0.10.1" }
 pyyaml = "^5.4.1"
 rich = {extras = ["jupyter"], version = "^12.0.0"}
 sqlalchemy_utils = "0.38.3"
@@ -79,13 +78,14 @@ python-jose = { extras = ["cryptography"], version = "~3.3.0", optional = true}
 fastapi-utils = { version = "~0.2.1", optional = true}
 orjson = { version = "~3.8.3", optional = true}
 
-# optional dependencies for stack recipes
-
 # Optional dependencies for project templates
 copier = { version = ">=7.0.0", optional = true }
 jinja2-time = { version = "^0.2.0", optional = true }
 black = { version = "^23.3.0", optional = true }
 ruff = { version = "^0.0.270", optional = true }
+
+# Optional terraform dependency for stack recipes and ZenServer deployments
+python-terraform = { version = "^0.10.1", optional = true }
 
 # Optional dependencies for the AWS secrets store
 boto3 = { version=">=1.16.0,<=1.24.59", optional = true }
@@ -166,6 +166,7 @@ docker-compose = { version = "^1.28", optional = true }
 [tool.poetry.extras]
 server = ["fastapi", "uvicorn", "python-multipart", "python-jose", "fastapi-utils", "orjson"]
 templates = ["copier", "jinja2-time", "black", "ruff"]
+terraform = ["python-terraform"]
 secrets-aws = ["boto3"]
 secrets-gcp = ["google-cloud-secret-manager"]
 secrets-azure = ["azure-identity", "azure-keyvault-secrets"]
