+++ src/zenml/cli/server.py
@@ -31,7 +31,7 @@ from zenml.console import console
 from zenml.enums import ServerProviderType, StoreType
 from zenml.exceptions import IllegalOperationError
 from zenml.logger import get_logger
-from zenml.utils import yaml_utils
+from zenml.utils import terraform_utils, yaml_utils
 from zenml.utils.analytics_utils import AnalyticsEvent, event_handler
 from zenml.zen_server.utils import get_active_deployment
 
@@ -368,6 +368,11 @@ def deploy(
         event=AnalyticsEvent.ZENML_SERVER_DEPLOYED,
         v2=True,
     ) as analytics_handler:
+        try:
+            terraform_utils.verify_terraform_installation()
+        except RuntimeError as e:
+            cli_utils.error(str(e))
+
         config_dict: Dict[str, Any] = {}
 
         if config:
