+++ src/zenml/utils/terraform_utils.py
@@ -1,0 +1,42 @@
+#  Copyright (c) ZenML GmbH 2023. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+"""Terraform utilities."""
+import zenml
+
+
+def verify_terraform_installation() -> None:
+    """Verifies the Terraform installation.
+
+    Raises:
+        RuntimeError: If Terraform is not installed or ZenML was installed
+            without the `terraform` extra.
+    """
+    try:
+        import python_terraform
+    except ImportError:
+        raise RuntimeError(
+            "Missing Terraform python library. This is probably because ZenML "
+            "was installed without the optional `terraform` extra. To install "
+            "the missing dependencies, run: \n\n"
+            f'`pip install "zenml[terraform]=={zenml.__version__}"`.'
+        )
+
+    try:
+        python_terraform.Terraform().cmd("-version")
+    except FileNotFoundError:
+        raise RuntimeError(
+            "Missing Terraform installation. Check out "
+            "https://developer.hashicorp.com/terraform/downloads for how to "
+            "install it on your operating system."
+        )
